Submission form & file upload/serve
The submission for is done-ish, doesn't actually put the data in the database yet, but it takes all the data and files properly.
Renamed TADs to be submissions, since that was the generalized name that we had agreed upon beforehand. 
Added major professor field to db and submission form.
updated links for dashboard and header